Decompiler: started work on opcode-to-script mappings
Wrote up some skeletal class definitions for OpcodeFuncToScriptMapping and OpcodeArgTypeinfo, along with some plans for how to finish them. We'll need these in order to implement decompiling in a relatively convenient way and in order to implement compiling at all.

+++ native/src/ReachVariantTool/game_variants/components/megalo/decompiler/decompiler.cpp
@@ -3,11 +3,11 @@
 namespace Megalo {
    void Decompiler::modify_indent_count(int16_t nesting) noexcept {
       if (nesting >= 0) {
-         size_t spaces = nesting * this->indent_size;
+         size_t spaces = (size_t)nesting * this->indent_size;
          for (size_t i = 0; i < spaces; i++)
             this->current_indent += ' ';
       } else {
-         size_t spaces = -nesting * this->indent_size;
+         size_t spaces = (size_t)(-nesting) * this->indent_size;
          size_t size   = this->current_indent.size();
          if (size < spaces)
             size = 0;
